[b-5-implement-sorting] review#2. 1) Genre changed to immutable. 2) added profile for separate running ITest
