+++ pom.xml
@@ -14,7 +14,7 @@
         <spring.version>5.1.3.RELEASE</spring.version>
         <postgresql.version>42.2.4</postgresql.version>
         <HikariCP.version>3.3.1</HikariCP.version>
-        <maven-compiler-plugin.version> 3.6.1</maven-compiler-plugin.version>
+        <maven-compiler-plugin.version>3.6.1</maven-compiler-plugin.version>
         <tomcat7-maven-plugin.version>2.2</tomcat7-maven-plugin.version>
         <jackson.version>2.9.9</jackson.version>
         <junit.version>4.12</junit.version>
@@ -24,6 +24,7 @@
         <logback-classic.version>1.2.3</logback-classic.version>
         <h2.version>1.4.199</h2.version>
         <lombok.version>1.18.8</lombok.version>
+        <maven-surefire-plugin.version>2.22.2</maven-surefire-plugin.version>
     </properties>
 
     <build>
@@ -48,8 +49,49 @@
                 </configuration>
             </plugin>
 
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>${maven-surefire-plugin.version}</version>
+                <configuration>
+                    <excludes>
+                        <exclude>**/*ITest</exclude>
+                    </excludes>
+                </configuration>
+            </plugin>
+
         </plugins>
     </build>
+
+    <profiles>
+        <profile>
+            <id>includeITest</id>
+            <build>
+                <plugins>
+                    <plugin>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <version>${maven-surefire-plugin.version}</version>
+                        <executions>
+                            <execution>
+                                <phase>integration-test</phase>
+                                <goals>
+                                    <goal>test</goal>
+                                </goals>
+                                <configuration>
+                                    <excludes>
+                                        <exclude>none</exclude>
+                                    </excludes>
+                                    <includes>
+                                        <include>**/*ITest</include>
+                                    </includes>
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
+    </profiles>
 
     <dependencies>
         <!--Spring -->
@@ -137,8 +179,6 @@
         </dependency>
 
     </dependencies>
-
-
 
 
 </project>
