+++ src/main/java/com/stroganova/movielandapp/dao/jdbc/mapper/GenreRowMapper.java
@@ -9,9 +9,8 @@ import java.sql.SQLException;
 public class GenreRowMapper implements RowMapper<Genre> {
     @Override
     public Genre mapRow(ResultSet resultSet, int i) throws SQLException {
-        Genre genre = new Genre();
-        genre.setId(resultSet.getLong("id"));
-        genre.setName(resultSet.getString("name"));
-        return genre;
+        long id = resultSet.getLong("id");
+        String name = resultSet.getString("name");
+        return new Genre(id, name);
     }
 }
