+++ src/main/java/com/stroganova/movielandapp/service/impl/DefaultGenreService.java
@@ -8,7 +8,6 @@ import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,14 +18,12 @@ import java.util.List;
 @FieldDefaults(makeFinal=true, level= AccessLevel.PRIVATE)
 public class DefaultGenreService implements GenreService {
 
-
     @NonNull
-    @Qualifier("cachedJdbcGenreDao")
-    GenreDao cachedJdbcGenreDao;
+    GenreDao genreDao;
 
     @Override
     public List<Genre> getAll() {
         log.info("Get all genres ");
-        return cachedJdbcGenreDao.getAll();
+        return genreDao.getAll();
     }
 }
