+++ src/main/resources/spring/rootContext.xml
@@ -29,7 +29,7 @@
         <property name="poolName" value="springHikariCP" />
         <property name="connectionTestQuery" value="SELECT 1" />
         <property name="dataSourceClassName" value="org.postgresql.ds.PGSimpleDataSource" />
-        <property name="maximumPoolSize" value="50" />
+        <property name="maximumPoolSize" value="10" />
         <property name="maxLifetime" value="60000" />
         <property name="idleTimeout" value="30000" />
         <property name="dataSourceProperties">
