+++ src/test/java/com/stroganova/movielandapp/dao/cache/CachedGenreDaoTest.groovy
@@ -1,22 +1,24 @@
-package com.stroganova.movielandapp.dao.jdbc
+package com.stroganova.movielandapp.dao.cache
 
+import com.stroganova.movielandapp.dao.jdbc.JdbcGenreDao
 import com.stroganova.movielandapp.entity.Genre
 import org.junit.Test
 
 import static org.mockito.Mockito.mock
 import static org.mockito.Mockito.when
 
-class CachedJdbcGenreDaoTest {
+class CachedGenreDaoTest {
 
     @Test
     void testGetAll() {
-        def expectedGenres = [new Genre(id: 1L, name: "genreFirst"),
-                              new Genre(id: 2L, name: "genreSecond")]
+
+        def expectedGenres = [new Genre(1L, "genreFirst"),
+                              new Genre(2L, "genreSecond")]
 
         def genreDao = mock(JdbcGenreDao.class)
         when(genreDao.getAll()).thenReturn(expectedGenres)
 
-        def cachedJdbcGenreDao = new CachedJdbcGenreDao(genreDao)
+        def cachedJdbcGenreDao = new CachedGenreDao(genreDao)
         cachedJdbcGenreDao.updateGenres()
         def actualGenres = cachedJdbcGenreDao.getAll()
